// src/app/shared/models/formType.ts
export enum FormElementsEnum {
  DynamicFormRequired = 'dynamicFormRequired',
  DynamicFormType = 'dynamicFormType',
  DynamicFormLabel = 'dynamicFormLabel',
  DynamicFormSelectOptions = 'dynamicFormSelectOptions',
  DynamicFormColSpan = 'dynamicFormColSpan',
}

export const options = ['Text', 'Number', 'Checkbox', 'Select', 'Email', 'SignaturePad'];